% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{adjustDriftLowess}
\alias{adjustDriftLowess}
\title{Adjust position based on lowess}
\usage{
adjustDriftLowess(xt, yi)
}
\arguments{
\item{xt}{}

\item{yi}{}
}
\value{
updated yi
}
\description{
Assume walking in straight line so x,y,z should all be straight
Assume mode is zero
}
