% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeFeatures.R
\name{appendRecordData}
\alias{appendRecordData}
\title{Appends health record information}
\usage{
appendRecordData(pframe, missingrecords)
}
\arguments{
\item{dframe}{}
}
\value{
update dframe
}
\description{
Appends health record information
}
