% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{buildFeatures}
\alias{buildFeatures}
\title{Build Features From Data Objects ... happens at "record-id" level (rv)}
\usage{
buildFeatures()
}
\value{
nothing
}
\description{
Build Features From Data Objects ... happens at "record-id" level (rv)
}
