% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitFeatures.R
\name{codeHealthState}
\alias{codeHealthState}
\title{Build variable $healthState = 0,1,2, 3,  5 and $isPD = 0, 0.5, 1  where 0.5 and 3 are NA.}
\usage{
codeHealthState(dframe)
}
\arguments{
\item{dframe}{}
}
\value{
updated dframe with $healthState and $isPD
}
\description{
Build variable $healthState = 0,1,2, 3,  5 and $isPD = 0, 0.5, 1  where 0.5 and 3 are NA.
}
\examples{
trainme = codeHealthState(mPower$walking.training);
}
