% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseData.R
\name{collapseTimeStampDeviceMotion}
\alias{collapseTimeStampDeviceMotion}
\title{Collapse Device Motion Timestamp}
\usage{
collapseTimeStampDeviceMotion(drecords)
}
\arguments{
\item{drecords}{dataframe of tlist [d]eviceMotion records}
}
\value{
updated dataframe
}
\description{
Apple timestamp for deviceMotion has only one timestamp.
This function flattens the file into panel form, with shared timestamp
}
