% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeFeatures.R
\name{compareRecords}
\alias{compareRecords}
\title{Title}
\usage{
compareRecords(submitr, auditr)
}
\arguments{
\item{submitr}{list from submit template, r as string}

\item{auditr}{list from harvest audit, rv as variable}
}
\value{
list (rv) of missing values (we don't have in audit)
}
\description{
Title
}
