% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateMe.R
\name{computeAngle}
\alias{computeAngle}
\title{Compute Angle between two vectors in n-dim space}
\usage{
computeAngle(v1, v2, out = "radians")
}
\arguments{
\item{v1}{numeric vector}

\item{v2}{numeric vector}

\item{out}{degree output format (default is 'radians', also 'degrees')}
}
\value{
numeric angle in 'out' format
}
\description{
https://stackoverflow.com/questions/1897704/angle-between-two-vectors-in-r
}
\examples{
v1 = c(2,1,0); v2 = c(0,1,2);
computeAngle(v1,v2);
computeAngle(v1,v2,out="degrees");

}
