% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{computeMotionDetails}
\alias{computeMotionDetails}
\title{Compute Motion Details}
\usage{
computeMotionDetails(rv, info, plotme = TRUE)
}
\arguments{
\item{rv}{record identifer, record variable (rv)}

\item{info}{contains motion-detected compressed forms of "outbound", "resting", "returnwalk"}

\item{plotme}{if true, we plot to PDF files in RECORDS folder}
}
\value{
list of details
}
\description{
Compute Motion Details
}
