% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{computeMotionWindow}
\alias{computeMotionWindow}
\title{Truncate data based on significant motion}
\usage{
computeMotionWindow(olist, act)
}
\arguments{
\item{olist}{list from orientToGravity}

\item{act}{should be "outbound" or "return"}
}
\value{
nlist truncated list with aligned timestamps based on motion
}
\description{
Truncate data based on significant motion
}
