% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{computeNormValue}
\alias{computeNormValue}
\title{Compute Range differences based on normalizing to min or max}
\usage{
computeNormValue(x)
}
\arguments{
\item{x}{}
}
\value{
list with details from computation
}
\description{
How far is "max" from the "min" in normalized units?
How far is "min" from the "max" in normalized units?
}
