% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUnits.R
\name{convertUnits}
\alias{convertUnits}
\title{Convert Units (from gravity to m/s^2)}
\usage{
convertUnits(x, from = "g", to = "m/s^2")
}
\arguments{
\item{x}{numeric vector}

\item{from}{units to change from}

\item{to}{units to change to}
}
\value{
numeric vector of converted result
}
\description{
Constants must be loaded with:  \code{setup = loadSetup();}
}
\examples{

convertUnits(1); # 9.8
convertUnits(seq(-2,2,by=0.25)); 


}
