% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitFeatures.R
\name{dampenOutliers}
\alias{dampenOutliers}
\title{Dampen outliers}
\usage{
dampenOutliers(dframe, xfeats, iCut = 6)
}
\arguments{
\item{dframe}{}

\item{xfeats}{numeric vector, list of columns to be evaluated}

\item{iCut}{number, number of IQRs to use to truncate min/max}
}
\value{
updated dframe
}
\description{
Dampen outliers
}
