% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{determineMotionInterval}
\alias{determineMotionInterval}
\title{Determine if the accelerometer is active, drop noise before full on walking}
\usage{
determineMotionInterval(dframe, mot = "pos", act = "outbound")
}
\arguments{
\item{dframe}{}
}
\value{
list of first and last index, with updated yn zero-centered acceleration with drift ...
}
\description{
Determine if the accelerometer is active, drop noise before full on walking
}
