% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseData.R
\name{determineOrder}
\alias{determineOrder}
\title{Determine Activity Order}
\usage{
determineOrder(mlist)
}
\arguments{
\item{mlist}{}
}
\value{
dataframe sorted by order with start/stop timestamps as values
}
\description{
[rest, outbound, return]
[outbound, rest, return]
[outbound, return, rest]
}
