% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestJSON.R
\name{doFileCopy}
\alias{doFileCopy}
\title{Copy Synapse Cache to Local Cache (internal private function)}
\usage{
doFileCopy(data, json, userFolder, json_files)
}
\arguments{
\item{data}{}

\item{json}{}

\item{userFolder}{}

\item{json_files}{}
}
\value{
integer 'localrecords'
}
\description{
Copy Synapse Cache to Local Cache (internal private function)
}
