% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateMe.R
\name{doIntegration}
\alias{doIntegration}
\title{doIntegration}
\usage{
doIntegration(xt, yi)
}
\arguments{
\item{xt}{}

\item{yi}{}
}
\value{
dataframe with time, acc, vel, pos
}
\description{
Input is xt, acc
}
\details{
Output is xt (time) with acc, vel, pos
Performs first and second integral
}
