% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{finalMotionFeatures}
\alias{finalMotionFeatures}
\title{Build Flatted Motion-Features dataframe}
\usage{
finalMotionFeatures(records, method = "string")
}
\arguments{
\item{records}{list of character strings of records.  From submission template, doesn't match audit 79168 - 79137}

\item{method}{is the records list "string" or "variable(s)"}
}
\value{
dataframe  features, then $r and $h
}
\description{
Build Flatted Motion-Features dataframe
}
