% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_h.R
\name{g_h}
\alias{g_h}
\title{Compute gravity at a certain altitude on earth.}
\usage{
g_h(h)
}
\arguments{
\item{h}{height in meters}
}
\value{
gravity on earth adjusted for altitude \code{h}
}
\description{
Constants must be loaded with:  \code{setup = loadSetup();}
}
\examples{
g10 = g_h(10);
g100 = g_h(100);
g1000 = g_h(1000);
}
