% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSetup.R
\name{getCompleteRecords}
\alias{getCompleteRecords}
\title{Get Complete Records}
\usage{
getCompleteRecords(dfstring)
}
\arguments{
\item{dfstring}{string representing a dataframe element in mPower list}
}
\value{
list of (rvs) - parseSingleRecordVariable(rv) will be evaluated to determine the JSON objects are valid
}
\description{
A complete record has 8 JSON objects.  We verifiy the JSON objects are valid as well.
}
\details{
Also stores an object of the badrecords
}
\examples{
mPower = loadSynapseData();	# loads summary info of data tables
setup = loadSetup();  	# builds constants
audit = harvestAudit();
recs = recordMap(mPower);
}
