% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestJSON.R
\name{getDemographics}
\alias{getDemographics}
\title{Append Demographics to audit}
\usage{
getDemographics(myH)
}
\arguments{
\item{myH}{hv identifier for healthcode}
}
\value{
row of dframe
}
\description{
Append Demographics to audit
}
