% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeFeatures.R
\name{getMotionFeatures}
\alias{getMotionFeatures}
\title{Get Features from Motion Data}
\usage{
getMotionFeatures(records, method = "string")
}
\arguments{
\item{records}{list of character strings of records.  From submission template, doesn't match audit 79168 - 79137}
}
\value{
dataframe of motion options
}
\description{
Get Features from Motion Data
}
