% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{getMotionFeaturesFromRecord}
\alias{getMotionFeaturesFromRecord}
\title{Motion Features}
\usage{
getMotionFeaturesFromRecord(rv, plotme = TRUE)
}
\arguments{
\item{rv}{record identifier in variable form [rv]}

\item{plotme}{if true, we plot to a PDF file in the folder "motion-details" of the RECORD}
}
\value{
list of features [recordObj]
}
\description{
Motion Features
}
