% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeFeatures.R
\name{getPedometerFeatures}
\alias{getPedometerFeatures}
\title{Get Features from Pedometer Data}
\usage{
getPedometerFeatures(records, method = "string")
}
\arguments{
\item{records}{list of character strings of records.  From submission template, doesn't match audit 79168 - 79137}
}
\value{
dataframe of pedometer options
}
\description{
Get Features from Pedometer Data
}
