% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{getPedometerFeaturesFromRecord}
\alias{getPedometerFeaturesFromRecord}
\title{Pedometer Features}
\usage{
getPedometerFeaturesFromRecord(rvObj)
}
\arguments{
\item{rvObj}{}
}
\value{
list of features
}
\description{
Pedometer Features
}
