% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateMe.R
\name{getQuaternion}
\alias{getQuaternion}
\title{Create quaternion (3-D space)}
\usage{
getQuaternion(v1, v2)
}
\arguments{
\item{v1}{numeric vector of length 3}

\item{v2}{numeric vector of length 3}
}
\value{
numeric vector of length 4 [w  x,y,z] with *norm 1*
}
\description{
Represent the transformations required to get from v1 to v2
}
\examples{
getQuaternion(v1,v2);

}
