% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordStringToVariable.R
\name{getRecordPath}
\alias{getRecordPath}
\title{Build Record Path from (rv)}
\usage{
getRecordPath(rv)
}
\arguments{
\item{rv}{}
}
\value{
string, recordFolder
}
\description{
requires audit$rclist
}
