% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseData.R
\name{getSecondsFromPedometerTime}
\alias{getSecondsFromPedometerTime}
\title{Parse Pedometer Data}
\usage{
getSecondsFromPedometerTime(precords)
}
\arguments{
\item{precords}{dataframe of raw [p]edometer records}
}
\value{
updated dataframe
}
\description{
Parse Pedometer Data
}
