% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestJSON.R
\name{getSingleHealthCodeQuery}
\alias{getSingleHealthCodeQuery}
\title{SQL query to Synapse to get all records for Single Health Code}
\usage{
getSingleHealthCodeQuery(h, syntable = "syn10146553")
}
\arguments{
\item{h}{}

\item{syntable}{}
}
\value{
Synapse custom 'table' template
}
\description{
SQL query to Synapse to get all records for Single Health Code
}
