% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestJSON.R
\name{getSingleRecordIdQuery}
\alias{getSingleRecordIdQuery}
\title{SQL query to Synapse to get Single Record}
\usage{
getSingleRecordIdQuery(h, r, syntable = "syn10146553")
}
\arguments{
\item{h}{}

\item{r}{}

\item{syntable}{}
}
\value{
Synapse custom 'table' template
}
\description{
SQL query to Synapse to get Single Record
}
