% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestJSON.R
\name{harvestAudit}
\alias{harvestAudit}
\title{Perform an Audit on localCache of objects}
\usage{
harvestAudit(path = "userObjects", key = "AUDIT")
}
\arguments{
\item{path}{string that can be relative to 'localCache' or absolute}

\item{key}{string used for caching a unique audit}
}
\value{
list named 'audit'
}
\description{
Good/Bad based on completeness of records
Lists organize the data on different keying features for quick lookups
}
\examples{
audit = harvestAudit();
}
