% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestJSON.R
\name{harvestJSON}
\alias{harvestJSON}
\title{Harvest JSON data-motion objects from Synapse}
\usage{
harvestJSON(dframe, syntable, random = T, verbose = T)
}
\arguments{
\item{dframe}{data.frame with a list of \code{healthCode} s.}

\item{syntable}{string, tablename to do query and download}

\item{random}{boolean (true or false); if true, we randomize the list (allows for multiple instances)}

\item{verbose}{boolean (true or false); if true, we print messages to console}
}
\value{
number of total records
}
\description{
Harvest JSON data-motion objects from Synapse
}
\examples{

harvestJSON(mPower$walking.training,     'syn10146553');
harvestJSON(mPower$walking.testing,      'syn10733842');
harvestJSON(mPower$walking.supplemental, 'syn10733835');
harvestJSON(mPower$walking.training,     'syn10733842',random=F,verbose=F);


}
