% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestJSON.R
\name{harvestSingle}
\alias{harvestSingle}
\title{Harvest JSON data-motion objects from Synapse (single healthCode)}
\usage{
harvestSingle(h, syntable = "syn10146553")
}
\arguments{
\item{h}{a string, the healthCode}

\item{syntable}{a string, the relevant table for SQL}
}
\value{
integer, the number of records
}
\description{
Harvest JSON data-motion objects from Synapse (single healthCode)
}
\examples{




}
