% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeFeatures.R
\name{imputateDataFrame}
\alias{imputateDataFrame}
\title{Imputate}
\usage{
imputateDataFrame(dframe, myC)
}
\arguments{
\item{dframe}{dataframe with NA, NAN, INF, -INF values}

\item{myC}{columns to imputate}
}
\value{
updated dataframe
}
\description{
We first cache all unique health code missing values.
We replace NAN/INF with NA
We loop once to replace health code groups with median values
We loop again to replace global median values
}
\details{
[TODO: address extreme outliers]
}
