% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseData.R
\name{initializeRecordTimeList}
\alias{initializeRecordTimeList}
\title{Initialize Time-flattened List}
\usage{
initializeRecordTimeList()
}
\value{
list 'tlist'
}
\description{
Initialize Time-flattened List
}
