% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateMe.R
\name{integrateMe}
\alias{integrateMe}
\title{Numerical Integration (x,y) \code{integrateMe}}
\usage{
integrateMe(x, y, method = "trapezoid")
}
\arguments{
\item{x}{a numeric vector representing x-axis}

\item{y}{a numeric vector representing y-axis, must be of same length as x}

\item{method}{a string representing the numerical integration approach 'trapezoid'}
}
\value{
list with "area", "dydx" cumulative integral from x[1] to current x[i].
}
\description{
Numerical Integration (x,y) \code{integrateMe}
}
\examples{
# Calculate the area under the sine curve from 0 to 2*pi:
n <- 202
x <- seq(0, 2*pi, len = n)
y <- sin(x)
ans <- integrateMe(x, y); #=> 0
str(ans); 
plot(x,y,type="l",col="black",xlab="",ylab="",xlim=c(0,2*pi),ylim=c(-1,2));
abline(h=0)
par(new=T);
plot(x,ans$dydx, col="green",xlab="",ylab="",xlim=c(0,2*pi),ylim=c(-1,2));
}
