% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSetup.R
\name{loadSetup}
\alias{loadSetup}
\title{Load Setup}
\usage{
loadSetup()
}
\value{
list 'setup' is returned.
}
\description{
Loads constant variables used in package
}
\examples{
setup = loadSetup();  	# builds constants
str(setup);

}
