% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitFeatures.R
\name{loadSubmitLibraries}
\alias{loadSubmitLibraries}
\title{Load required libraries for Synapse model prediction}
\usage{
loadSubmitLibraries()
}
\value{
nothing
}
\description{
We assume and
}
\details{
library(synapseClient);  
synapseLogin();

have already been called.
}
