% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSynapse.R
\name{loadSynapse}
\alias{loadSynapse}
\title{Load Synapse Library and Login to Access the Data}
\usage{
loadSynapse(configfile, login = TRUE, rememberMe = TRUE,
  synapseUserOverride = character(0), synapsePassOverride = character(0))
}
\arguments{
\item{configfile}{full path to configuration file with necessary instance variables (username, password, localcache)}

\item{login}{boolean value (true or false) to process login with}

\item{rememberMe}{boolean value (true or false) to keep session alive with Synapse}

\item{synapseUserOverride}{string to override the config-file username to login to Synapse}

\item{synapsePassOverride}{string to override the config-file password to login to Synapse}
}
\value{
string on success
}
\description{
The config file has username, password, and local cache folders
}
\examples{
loadSynapse('./mPowerEI/example/config.txt');

loadSynapse('./mPowerEI/example/config.txt',login=F);

loadSynapse('./mPowerEI/example/config.txt',login=T,synapseUserOverride='monte.shaffer@gmail.com',synapsePassOverride='myNewPassword');

}
