% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSynapseData.R
\name{loadSynapseData}
\alias{loadSynapseData}
\title{Load Synapse Data From Tables Assigned to Variables}
\usage{
loadSynapseData(synapseTablesOverride = NULL,
  synapseVariablesOverride = NULL)
}
\arguments{
\item{synapseTablesOverride}{[optional] string vector to override the config-file synapseTables}

\item{synapseVariablesOverride}{[optional] string vector to override the config-file synapseVariables}
}
\value{
vector of strings on success
}
\description{
The config file has default values: synapseTables and synapseVariables should be of the same length
}
\examples{
loadSynapse('./mPowerEI/example/config.txt'); loadSynapseData();
 
}
