% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordStringToVariable.R
\name{md5}
\alias{md5}
\title{MD5 Encryption}
\usage{
md5(txt)
}
\arguments{
\item{txt}{a text string or a vector of text strings}
}
\value{
a text string or a vector of text strings
}
\description{
One-way encyption using md5 methodology from library digest::digest
}
\examples{

md5('hello-again/#&0!_ugly-characters-break-R');

md5(c('hello','again'));

}
