% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{mergeExtremes}
\alias{mergeExtremes}
\title{Merge global extremes based on one-two constraint}
\usage{
mergeExtremes(subMin, subMax, subMinE, subMaxE, animate = FALSE)
}
\arguments{
\item{subMin}{}

\item{subMax}{}

\item{subMinE}{}

\item{subMaxE}{}

\item{animate}{if true, we plot and animate the results (e.g., debugging)}
}
\value{
list with updated values
}
\description{
Merge global extremes based on one-two constraint
}
