% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseData.R
\name{mergeListsAccelDeviceMotion}
\alias{mergeListsAccelDeviceMotion}
\title{Combine Independent Data (accel and deviceMotion)}
\usage{
mergeListsAccelDeviceMotion(slist)
}
\arguments{
\item{slist}{a data list with the timestamps smoothed to a setup$designpoint (e.g., 100)}
}
\value{
mlist an updated data list with each named element in flattened-panel form.
}
\description{
With common time, we can now merge the data sets into a flattened panel
}
