% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{Mode}
\alias{Mode}
\title{Compute mode}
\usage{
Mode(x)
}
\arguments{
\item{x}{}
}
\value{
mode
}
\description{
https://stackoverflow.com/questions/2547402/is-there-a-built-in-function-for-finding-the-mode
}
