% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitFeatures.R
\name{mPower_test}
\alias{mPower_test}
\title{Internal Test of Greedy Ensemble with Predictions and Accuracy}
\usage{
mPower_test(submitme, dv = "healthState", dvv = c(0, 5), dvf = c("PD.off",
  "nonPD"), xfeats = 1:20, split = 80, dreplace = "none")
}
\arguments{
\item{submitme}{the dataframe with submission, not entire synapseSubmission, $r and $h are required.}

\item{dv}{string, the name of the column that will be the dv}

\item{dvv}{numeric vector, the values of the DV that will be classified, first = 0, second = 5}

\item{dvf}{string vector, the factors of the DV that will be classified, first = 'PD.off', second = 'nonPD'}

\item{xfeats}{numeric vector, the column numbers that define the features}

\item{split}{number, 1-99, what percent of the data will be sampled into train vs test set (e.g., 80 means 80 into train, 20 into 'out of sample' test)}

\item{dreplace}{string, how do we replace duplicates in analysis set, "first" is just the first value, "median" is the median of all values, "none" skips this step...}
}
\value{
list of items from caret, model = , ensemble = , predict = , accuracy = , control =
}
\description{
Internal Test of Greedy Ensemble with Predictions and Accuracy
}
