% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseData.R
\name{orientToGravity}
\alias{orientToGravity}
\title{Orient To Gravity}
\usage{
orientToGravity(mlist)
}
\arguments{
\item{mlist}{}
}
\value{
list 'olist' of mlist with updated orientation
}
\description{
The deviceMotion 'gravity' data is compared to normal downward gravity.
gravityDown = c(0,0,-1);
The getRotationMatrix is determined between g' and g.
This matrix is used to alter the single data points.
}
\details{
This calculation occurs at every designpoint.
Simultaneously, within this function, an angle is computed between each point.
}
