% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseData.R
\name{parseAllRecordData}
\alias{parseAllRecordData}
\title{parseAllRecordData}
\usage{
parseAllRecordData(verbose = TRUE, force = FALSE)
}
\arguments{
\item{verbose}{boolean; if TRUE, will print progress to console.}

\item{force}{boolean; if TRUE, it will force a rebuild of all objects}
}
\value{
nothing
}
\description{
parseAllRecordData
}
