% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseData.R
\name{parseSingleRecordVariable}
\alias{parseSingleRecordVariable}
\title{Parse a Single Record (rv)}
\usage{
parseSingleRecordVariable(rv, force = FALSE)
}
\arguments{
\item{rv}{a string of the record in variable form "RECORD28832018b9c849e79d67f201d99eb8b6"}

\item{force}{boolean; if TRUE, it will force a rebuild of all objects}
}
\value{
list of two objects 'raw' and 'motion':  raw is the raw JSON data, motion is the data redesigned based on setup$designpoint
}
\description{
Parse a Single Record (rv)
}
