% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseData.R
\name{rebuildTimeIncrement}
\alias{rebuildTimeIncrement}
\title{Rebuild Data to Time Design Point}
\usage{
rebuildTimeIncrement(temp, increment)
}
\arguments{
\item{temp}{dataframe with first column timestamp, all other columns numeric}

\item{increment}{designpoint frame (determined in setup$designpoint)}
}
\value{
updated dataframe with few points, and median values
}
\description{
We use a recent median up to a right-hand stop point.
setup$pareto determines if the point will be included.
[TODO] - missing data inline ... do not know if it is possible, but maybe control
}
