% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSetup.R
\name{recordMap}
\alias{recordMap}
\title{Map table data to Records (rv)}
\usage{
recordMap(mPower)
}
\arguments{
\item{mPower}{list of data objects}
}
\value{
lists 'recs' of data objects mapped on recordStringToVariable(RecordId)
}
\description{
Map table data to Records (rv)
}
\examples{
mPower = loadSynapseData();	# loads summary info of data tables
recs = recordMap(mPower); # organizes data based on rv
}
