% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recordStringToVariable.R
\name{recordVariableToString}
\alias{recordVariableToString}
\title{Convert recordId to a variable that can be used in R}
\usage{
recordVariableToString(r_, prepend = "RECORD")
}
\arguments{
\item{r_}{a record variable, string in form 'RECORD224f8562c92340feb29ea6af1cc7ca73'}

\item{prepend}{a string to 'prepend' to the updated string}
}
\value{
a string record string in form '224f8562-c923-40fe-b29e-a6af1cc7ca73'
}
\description{
Uses implode from library(tractor.base)
}
\examples{
r='224f8562-c923-40fe-b29e-a6af1cc7ca73'
rv = recordStringToVariable(r)
rvr = recordVariableToString(rv)
r==rvr

h='7457113a-34bd-4324-935c-f66bd33f1e4b';
hv = recordStringToVariable(h,prepend="HEALTH");
hvh = recordVariableToString(hv,prepend="HEALTH"); 


}
