% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitFeatures.R
\name{replaceDuplicates}
\alias{replaceDuplicates}
\title{Replace 'healthCode' duplicates in dataframe}
\usage{
replaceDuplicates(dframe, xpos, dreplace = "first")
}
\arguments{
\item{dframe}{dataframe}

\item{xpos}{positions of xfeats that need to be analyzed}

\item{dreplace}{string, how to replace with "first" value or "median"}
}
\value{
updated dataframe without replicates
}
\description{
Replace 'healthCode' duplicates in dataframe
}
