% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseData.R
\name{scaleToTimeIncrement}
\alias{scaleToTimeIncrement}
\title{Scale Data to Design Point}
\usage{
scaleToTimeIncrement(tlist, increment)
}
\arguments{
\item{tlist}{}

\item{increment}{(in milliseconds) assigned as setup$designpoint}
}
\value{
updated list 'slist'
}
\description{
Scale Data to Design Point
}
