% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitFeatures.R
\name{scoreFactor}
\alias{scoreFactor}
\title{Convert Factor to Number}
\usage{
scoreFactor(dframerow, myK, myS)
}
\arguments{
\item{dframerow}{}

\item{myK}{}

\item{myS}{}
}
\value{
single numeric value
}
\description{
Convert Factor to Number
}
