% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitFeatures.R
\name{stepwiseFeatureSelection}
\alias{stepwiseFeatureSelection}
\title{Select Best Features [Step-wise] to maximize ROC}
\usage{
stepwiseFeatureSelection(dframe, xfeats, rnum, fastignore = TRUE,
  ignoreme = c(0, 0), startfeats = NULL)
}
\arguments{
\item{dframe}{dataframe}

\item{xfeats}{columns with features}

\item{rnum}{number, column index or $r}

\item{fastignore}{if true, we will drop any negative ROC deltas}

\item{ignoreme}{if fastignore=true, these are the values to drop default 0,0 ... the first is for the first pass, the latter is for the second pass.  by default negative ROC deltas}

\item{startfeats}{a numeric vector to start... startfeats = c(151,13,76) ... this would predefine the first set of features, enables restarts ... [we have also updated the Rerror.txt to cache individual 'unique' ensemble calls]}
}
\value{
list of results, roc.nest is the xfeats in order that matter
}
\description{
Hierarchy:  pick single feature with highest ROC, include in model; include second-best feature (with first feature) with highest ROC, include in model; and so on.
}
\details{
We stop when the next feature to be added lowers the overall ROC.

setup multi-cores before call, if possible
library(doMC);
registerDoMC(cores = 16);

nested loop with stop
find first feature with highest roc, place that in the system, loop by adding all remaining features, add one at a time until ROC isn't improving...
default iCut

dframe = pfeats; xfeats=1:20; rnum=22;
}
