% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapseSubmit.R
\name{synapse.download_merge_covariate}
\alias{synapse.download_merge_covariate}
\title{Covariate data downloaded remotely from syntable}
\usage{
synapse.download_merge_covariate(features)
}
\arguments{
\item{features}{}
}
\value{
mergeddata with covariates (age/gender) from demos$
}
\description{
Covariate data downloaded remotely from syntable
}
