% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapseSubmit.R
\name{synapse.fit_model}
\alias{synapse.fit_model}
\title{Ensemble Training}
\usage{
synapse.fit_model(training, featurenames, covs_num, covs_fac)
}
\arguments{
\item{training}{}

\item{featurenames}{}

\item{covs_num}{}

\item{covs_fac}{}
}
\value{
greedy_ensemble
}
\description{
Ensemble Training
}
