% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapseSubmit.R
\name{synapse.loadSubmitLibraries}
\alias{synapse.loadSubmitLibraries}
\title{Load required libraries for Synapse model prediction}
\usage{
synapse.loadSubmitLibraries()
}
\value{
nothing
}
\description{
We assume and
}
\details{
library(synapseClient);  
synapseLogin();

have already been called.
}
