% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapseSubmit.R
\name{synapse.PD_score_challenge1}
\alias{synapse.PD_score_challenge1}
\title{mPower Challenge scoring}
\usage{
synapse.PD_score_challenge1(training_features)
}
\arguments{
\item{training_features}{}
}
\value{
ensemble_model
}
\description{
resultme = PD_score_challenge1(submitme);
str(resultme$error);
}
