% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateMe.R
\name{unitVector}
\alias{unitVector}
\title{Normalize a vector so the magnitude is 1 (unit radius)}
\usage{
unitVector(v)
}
\arguments{
\item{v}{numeric vector of coordinates}
}
\value{
numeric vector of coordinates (norm 1)
}
\description{
Normalize a vector so the magnitude is 1 (unit radius)
}
\examples{
unitVector(v1)
unitVector(v2)
}
