% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateMe.R
\name{vectorMagnitude}
\alias{vectorMagnitude}
\title{Determine length or magnitude of n-dim vector}
\usage{
vectorMagnitude(v)
}
\arguments{
\item{v}{numeric vector of coordinates}
}
\value{
numeric magnitude
}
\description{
Determine length or magnitude of n-dim vector
}
\examples{
vectorMagnitude(v1);
vectorMagnitude(v1-v2);

}
