% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFeatures.R
\name{getMotionFeaturesFromRecord}
\alias{getMotionFeaturesFromRecord}
\title{Motion Features}
\usage{
getMotionFeaturesFromRecord(rvObj)
}
\arguments{
\item{rvObj}{}
}
\value{
list of features
}
\description{
Motion Features
}
