% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitFeatures.R
\name{stepwiseFeatureSelection}
\alias{stepwiseFeatureSelection}
\title{Select Best Features [Step-wise] to maximize ROC}
\usage{
stepwiseFeatureSelection(dframe, xfeats, rnum)
}
\arguments{
\item{dframe}{dataframe}

\item{xfeats}{columns with features}

\item{rnum}{number, column index or $r}
}
\value{
list of results, roc.nest is the xfeats in order that matter
}
\description{
Hierarchy:  pick single feature with highest ROC, include in model; include second-best feature (with first feature) with highest ROC, include in model; and so on.
}
\details{
We stop when the next feature to be added lowers the overall ROC.

setup multi-cores before call, if possible
library(doMC);
registerDoMC(cores = 16);

nested loop with stop
find first feature with highest roc, place that in the system, loop by adding all remaining features, add one at a time until ROC isn't improving...
default iCut

dframe = pfeats; xfeats=1:20; rnum=22;
}
